Organise validation package
