+++ src/main/java/com/hedera/cli/hedera/crypto/CryptoTransfer.java
@@ -11,6 +11,9 @@ import com.fasterxml.jackson.databind.ObjectWriter;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.cli.config.InputReader;
 import com.hedera.cli.hedera.Hedera;
+import com.hedera.cli.hedera.validation.ValidateAccounts;
+import com.hedera.cli.hedera.validation.ValidateAmount;
+import com.hedera.cli.hedera.validation.ValidateTransferList;
 import com.hedera.cli.models.AccountManager;
 import com.hedera.cli.models.PreviewTransferList;
 import com.hedera.cli.models.TransactionManager;
